+++ applications/rtl_xmc4_irq/config.h
@@ -26,7 +26,7 @@
 #define MBX1_smc_b  0x22
 
 #define SM2_sma     0x1100
-#define SM2_smc     0x24
+#define SM2_smc     0x64
 #define SM2_act     1
 #define SM3_sma     0x1180
 #define SM3_smc     0x20
