+++ applications/rtl_xmc4_irq/slave_objectlist.c
@@ -3,6 +3,14 @@
 #include "utypes.h"
 #include <stddef.h>
 
+#ifndef HW_REV
+#define HW_REV "1.0"
+#endif
+
+#ifndef SW_REV
+#define SW_REV "1.0"
+#endif
+
 static const char acName1000[] = "Device Type";
 static const char acName1000_0[] = "Device Type";
 static const char acName1008[] = "Device Name";
@@ -90,17 +98,17 @@ const _objd SDO1008[] =
 };
 const _objd SDO1009[] =
 {
-  {0x0, DTYPE_VISIBLE_STRING, 24, ATYPE_RO, acName1009_0, 0, "1.0"},
+  {0x0, DTYPE_VISIBLE_STRING, 24, ATYPE_RO, acName1009_0, 0, HW_REV},
 };
 const _objd SDO100A[] =
 {
-  {0x0, DTYPE_VISIBLE_STRING, 24, ATYPE_RO, acName100A_0, 0, "1.0"},
+  {0x0, DTYPE_VISIBLE_STRING, 24, ATYPE_RO, acName100A_0, 0, SW_REV},
 };
 const _objd SDO1018[] =
 {
   {0x00, DTYPE_UNSIGNED8, 8, ATYPE_RO, acName1018_00, 4, NULL},
   {0x01, DTYPE_UNSIGNED32, 32, ATYPE_RO, acName1018_01, 0x1337, NULL},
-  {0x02, DTYPE_UNSIGNED32, 32, ATYPE_RO, acName1018_02, 4300, NULL},
+  {0x02, DTYPE_UNSIGNED32, 32, ATYPE_RO, acName1018_02, 0x4300, NULL},
   {0x03, DTYPE_UNSIGNED32, 32, ATYPE_RO, acName1018_03, 0, NULL},
   {0x04, DTYPE_UNSIGNED32, 32, ATYPE_RO, acName1018_04, 0x00000000, NULL},
 };
